requirements.txt
python
os
PIL
tkinter
pathlib
pycryptodome